[MINOR] code refactor Bookie env check during start.
move bookie environment check logic code to `BookieEnvirnmentChecker`
this may makes code much clearer.

no logic change or method delete.
